test release

+++ lib/database/src/Car.cpp
@@ -3,7 +3,7 @@
 Car::Car(uint64_t car_id,
          std::string name,
          uint64_t price,
-         uint64_t consumption,
+         double consumption,
          uint64_t capacity,
          std::string fuel,
          std::string picture_path,
